# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Chamados",
        "description": "Documentação da API de Chamados",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/centros_de_custo": {
            "get": {
                "tags": [
                    "Centros de Custo"
                ],
                "summary": "Lista todos os centros de custo",
                "operationId": "17581cf75260e9efbbea4d018d7b5155",
                "responses": {
                    "200": {
                        "description": "Lista de centros de custo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CentroDeCusto"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Centros de Custo"
                ],
                "summary": "Cria um novo centro de custo",
                "operationId": "e4e823d08ca2c2ec2d6f7860c6a03fe9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao",
                                    "codigo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Nome do Centro de Custo"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição do Centro de Custo"
                                    },
                                    "codigo": {
                                        "type": "string",
                                        "example": "001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Centro de custo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CentroDeCusto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/centros_de_custo/{id}": {
            "get": {
                "tags": [
                    "Centros de Custo"
                ],
                "summary": "Mostra um centro de custo específico",
                "operationId": "ef3ce3526ce410534267cc8dd4dfc4d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do centro de custo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do centro de custo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CentroDeCusto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centro de custo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Centros de Custo"
                ],
                "summary": "Atualiza um centro de custo existente",
                "operationId": "50a11089f91fc731efe0167728b12d6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do centro de custo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao",
                                    "codigo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Nome do Centro de Custo"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição do Centro de Custo"
                                    },
                                    "codigo": {
                                        "type": "string",
                                        "example": "001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Centro de custo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CentroDeCusto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centro de custo não encontrado"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Centros de Custo"
                ],
                "summary": "Deleta um centro de custo",
                "operationId": "244d4e45ebf63f148ca036de38722f98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do centro de custo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Centro de custo deletado com sucesso"
                    },
                    "404": {
                        "description": "Centro de custo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CentroDeCusto": {
                "required": [
                    "nome",
                    "descricao",
                    "codigo"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "nome": {
                        "type": "string"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "codigo": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}